+++ presto-main/src/main/java/com/facebook/presto/operator/ExchangeClient.java
@@ -20,6 +20,7 @@ import com.facebook.presto.execution.buffer.SerializedPage;
 import com.facebook.presto.memory.context.LocalMemoryContext;
 import com.facebook.presto.operator.PageBufferClient.ClientCallback;
 import com.facebook.presto.operator.WorkProcessor.ProcessState;
+import com.facebook.presto.spi.PrestoException;
 import com.google.common.collect.ImmutableList;
 import com.google.common.util.concurrent.Futures;
 import com.google.common.util.concurrent.ListenableFuture;
@@ -37,6 +38,7 @@ import java.util.ArrayList;
 import java.util.Deque;
 import java.util.LinkedList;
 import java.util.List;
+import java.util.Locale;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
@@ -46,6 +48,7 @@ import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicReference;
 
+import static com.facebook.presto.spi.StandardErrorCode.GENERIC_INTERNAL_ERROR;
 import static com.facebook.presto.spi.block.PageBuilderStatus.DEFAULT_MAX_PAGE_SIZE_IN_BYTES;
 import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Preconditions.checkState;
@@ -185,8 +188,17 @@ public class ExchangeClient
 
         checkState(!noMoreLocations, "No more locations already set");
 
+        RpcShuffleClient resultClient;
+        if (location.getScheme().toLowerCase(Locale.ENGLISH).startsWith("http")) {
+            resultClient = new HttpRpcShuffleClient(httpClient, location);
+        }
+        else {
+            // TODO: create thrift client based on scheme
+            throw new PrestoException(GENERIC_INTERNAL_ERROR, "unsupported task result client scheme " + location.getScheme());
+        }
+
         PageBufferClient client = new PageBufferClient(
-                httpClient,
+                resultClient,
                 maxErrorDuration,
                 acknowledgePages,
                 location,
